MAINT: renamed spmd._common.py to spmd._base.py for BaseEstimatorSPMD
