+++ onedal/spmd/covariance/covariance.py
@@ -17,7 +17,7 @@
 from onedal.covariance import EmpiricalCovariance as EmpiricalCovariance_Batch
 
 from ..._device_offload import support_usm_ndarray
-from .._common import BaseEstimatorSPMD
+from .._base import BaseEstimatorSPMD
 
 
 class EmpiricalCovariance(BaseEstimatorSPMD, EmpiricalCovariance_Batch):
